+++ CS500-framework/raytrace.cpp
@@ -85,7 +85,7 @@ void Scene::Command(const std::string c, const std::vector<double> f, const std:
         // Eye position (x,y,z),  view orientation (qw qx qy qz),  frustum height ratio ry
         if (m_isRealTime)
         {
-            m_RealTime->setCamera(glm::vec3(f[0], f[1], f[2]), glm::quat(f[3], f[4], f[5], f[6]), f[7]);
+            m_RealTime->setCamera(glm::vec3(f[0], f[1], f[2]), glm::quat((float)f[3], (float)f[4], (float)f[5], (float)f[6]), (float)f[7]);
         }
         else
         {
@@ -116,7 +116,7 @@ void Scene::Command(const std::string c, const std::vector<double> f, const std:
         // Creates a Material instance to be picked up by successive shapes
         if (m_isRealTime)
         {
-            m_RealTime->brdf(glm::vec3(f[0], f[1], f[2]), glm::vec3(f[3], f[4], f[5]), f[6]);
+            m_RealTime->brdf(glm::vec3(f[0], f[1], f[2]), glm::vec3(f[3], f[4], f[5]), (float)f[6]);
         }
         else
         {
@@ -163,7 +163,7 @@ void Scene::Command(const std::string c, const std::vector<double> f, const std:
         // Creates a Shape instance for a sphere defined by a center and radius
         if (m_isRealTime)
         {
-            m_RealTime->sphere(glm::vec3(f[0], f[1], f[2]), f[3]);
+            m_RealTime->sphere(glm::vec3(f[0], f[1], f[2]), (float)f[3]);
         }
         else
         {
@@ -191,7 +191,7 @@ void Scene::Command(const std::string c, const std::vector<double> f, const std:
         // Creates a Shape instance for a cylinder defined by a base point, axis vector, and radius
         if (m_isRealTime)
         {
-            m_RealTime->cylinder(glm::vec3(f[0], f[1], f[2]), glm::vec3(f[3], f[4], f[5]), f[6]);
+            m_RealTime->cylinder(glm::vec3(f[0], f[1], f[2]), glm::vec3(f[3], f[4], f[5]), (float)f[6]);
         }
         else
         {
@@ -281,8 +281,8 @@ void Scene::Command(const std::string c, const std::vector<double> f, const std:
         //   45 Z --> (0.924 0.000 0.000 0.383) --> (0.653 0.653 0.271 0.271)
         quat q, accum(1,0,0,0);
         for (int c=0;  c<f.size();  c+=2) {
-            float w = cos(f[c]*PI/180.0/2.0); // Cos and sin of angle/2 (as required by quaternions)
-            float d = sin(f[c]*PI/180.0/2.0);
+            float w = (float)cos(f[c]*PI/180.0/2.0); // Cos and sin of angle/2 (as required by quaternions)
+            float d = (float)sin(f[c]*PI/180.0/2.0);
             int axis = int(f[c+1]);    // Axis 1, 2, 3 (for X, Y, Z)
             #define C(a) ((axis==a) ? 1.0 : 0.0)
             q = quat(w, d*C(1), d*C(2), d*C(3));
